requirements.txt
# Core dependencies
torch>=1.12.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Fairness and explainability
fairlearn>=0.7.0
shap>=0.41.0
aif360>=0.5.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Configuration
pyyaml>=5.4.0

# Adversarial attacks
adversarial-robustness-toolbox>=1.10.0

# Development
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950

# Logging
tensorboard>=2.8.0
wandb>=0.12.0

# Utils
tqdm>=4.62.0